requirements.txt
httpx~=0.25.2
fastapi~=0.105.0
Pillow~=10.1.0